m2dsia_web_app/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# Logging
pyyaml==6.0.1

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
python-multipart==0.0.6

# Optional: For production
gunicorn==21.2.0